requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.0.7
django-paypal==2.1
gunicorn==23.0.0
idna==3.10
packaging==24.1
pillow==10.4.0
pytz==2024.2
requests==2.32.3
sqlparse==0.5.1
stripe==10.11.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
whitenoise==6.7.0
